backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
PyJWT==2.12.1
python-multipart
